Player dashboard implemented.

+++ src/main/java/com/tus/tpt/dto/session/CreateNewTrainingSession.java
@@ -3,5 +3,5 @@ package com.tus.tpt.dto.session;
 public record CreateNewTrainingSession(
         String datetime,
         String type,
-        long duration
+        int duration
 ) {}
